python/requirements.txt
pandas
tqdm
halo
matplotlib
plotly
openpyxl
requests
python-dotenv
